requirements.txt
torch
torchvision
matplotlib
requests
Pillow
